refactor: UserMenu, CheckNicknameButton src/contents 이하로 이동
